Flight Deals Notifier/requirements.txt
requests>=2.28.0
twilio>=8.0.0

python-dotenv~=1.2.1